Move WindowsConsole to its own file
